port(qcommon): bring 6 cm_*.cpp collision files into POSIX + Switch builds
The collision (cm_*) subsystem now links on both targets. 21 upstream
files total in the build.

Stub pass in posix_stubs.cpp added 24 symbols required by the cm_*
files:

REAL implementations (will collide with com_math.cpp when that ports):
- Vec3 family: Vec3Length, Vec3Add, Vec3Cross, Vec3Scale, Vec3Lerp,
  Vec3MadMad, Vec3IsNormalized
- Q_fabs, VecNCompareCustomEpsilon
- Matrix: MatrixTransformVector, MatrixTransposeTransformVector,
  G_TransposeMatrix, G_RotatePoint
- Plane geometry: PlaneFromPoints, IntersectPlanes (Cramer's rule),
  SnapPointToIntersectingPlanes

Engine stubs (replaced when their owning subsystems port):
- Hunk_Alloc: backed by new[]; leaks intentionally (matches upstream
  hunk lifetime).
- Sys_Error: same fputs+abort pattern as Com_Error.
- DB_FindXAssetHeader: returns default-constructed XAssetHeader (null).
- XModelTraceLine: clears trace_t (256-byte conservative memset).
- CM_LoadMapData_LoadObj: no-op.
- track_static_alloc_internal: no-op.

Globals:
- g_traceThreadInfo[THREAD_CONTEXT_COUNT]: 16 KB-per-slot backing storage
  via reinterpret_cast; the cm_* code in our build path only writes
  thread-local copies, never reads the array directly.
- useFastFile: nullptr.

Header fixes for the GCC build path (Switch):
- src/qcommon/cm_showcollision.cpp: cast (int)brush->numsides in for-loop
  bound for sign-compare safety.
- src/qcommon/cm_trace.cpp: 3 memset(trace_t) sites cast to (void*) so
  -Wclass-memaccess accepts the byte-blit (trace_t has constructors).
- src/gfx_d3d/r_sky.h, src/sound/snd_public.h, src/game/game_public.h,
  deps/ode/collision.h: [[maybe_unused]] on the static const char* name
  tables that GCC flags as unused at TU scope.

+++ src/qcommon/cm_showcollision.cpp
@@ -71,7 +71,7 @@ void __cdecl CM_ShowSingleBrushCollision(
                 drawCollisionPoly(*(unsigned int *)windingPool, (float (*)[3]) & windingPool[4], color);
             }
         }
-        for (sideIndex = 6; sideIndex < brush->numsides + 6; ++sideIndex)
+        for (sideIndex = 6; sideIndex < (int)brush->numsides + 6; ++sideIndex)
         {
             if (CM_BuildBrushWindingForSide(
                 (winding_t *)windingPool,
